requirements.txt
telethon>=1.34.0
pandas>=2.0.0
openpyxl>=3.1.0
python-telegram-bot>=20.0
